+++ backend/src/controllers/reports/Tab_14_rpt_P_BudgetSummary.js
@@ -1,6 +1,8 @@
 const useCommonComponents = require("../useCommonComponents/index");
 const useController = require("../useController/index");
-const model = require("@models/reports/Tab_14_rpt_P_BudgetSummary");
+//const model = require("@models/reports/Tab_14_rpt_P_BudgetSummary");
+const model = require("@models/reports/index");
+const projectModel = require("@models/projects");
 const utils = require("./helpers");
 const what = { single: "report", plural: "reports" };
 const controller = useController(model, what);
@@ -22,8 +24,15 @@ controller.Tab_14_rpt_P_BudgetSummary = async (request, reply) => {
   try {
     // Get the data from the database.
     const getDate = async () => new Date();
-
+    const projectId = Number(request.query.project);
     const result = {
+      project: await projectModel.findById(projectId),
+      budget: await model.getProjectBudget(projectId),
+      status: await projectModel.findMostRecentStatusById(projectId),
+      deliverable_summaries: await model.getDeliverableSummaries(projectId),
+      change_request: await model.getChangeRequests(projectId),
+      contracts: await model.getContracts(projectId),
+      contract_summaries: await model.getContractSummary(projectId),
       report_date: await getDate(),
     };