+++ backend/src/controllers/reports/Tab_15_rpt_P_QuarterlyBillingRequest.js
@@ -1,6 +1,7 @@
 const useCommonComponents = require("../useCommonComponents/index");
 const useController = require("../useController/index");
-const model = require("@models/reports/Tab_15_rpt_P_QuarterlyBillingRequest");
+//const model = require("@models/reports/Tab_15_rpt_P_QuarterlyBillingRequest");
+const model = require("@models/reports/index");
 const utils = require("./helpers");
 const what = { single: "report", plural: "reports" };
 const controller = useController(model, what);
@@ -22,11 +23,20 @@ controller.Tab_15_rpt_P_QuarterlyBillingRequest = async (request, reply) => {
   try {
     // Get the data from the database.
     const getDate = async () => new Date();
-
+    const projectId = Number(request.query.project);
+    const fiscal = Number(request.query.fiscal);
+    const quarter = Number(request.query.quarter);
     const result = {
+      project: await model.getProjectById(projectId),
+      deliverables: await model.getDeliverableBudgets(projectId, fiscal, quarter),
+      jv: await model.getJournalVoucher(projectId, fiscal, quarter),
+      client: await model.getClientCoding(projectId),
+      quarter: "Q" + quarter,
       report_date: await getDate(),
     };
 
+    // Calculate grand total from each deliverable amount.
+    result.deliverables_total = result.deliverables.reduce((acc, d) => acc + d.amount, 0);
     const body = await getDocumentApiBody(result, "Tab_15_rpt_P_QuarterlyBillingRequest.docx");
     const pdf = await cdogs.api.post("/template/render", body, pdfConfig);