+++ backend/src/controllers/reports/index.js
@@ -1,7 +1,6 @@
 const useController = require("../useController/index");
 const useCommonComponents = require("../useCommonComponents/index");
 const model = require("@models/reports/index");
-const projectModel = require("@models/projects");
 const what = { single: "report", plural: "reports" };
 const controller = useController(model, what);
 // Template and data reading
@@ -64,50 +63,6 @@ const getDocumentApiBody = async (
  * @param   {FastifyReply}   reply   FastifyReply is an instance of the standard http or http2 reply types.
  * @returns {object}
  */
-controller.getProjectBudgetReportOnRequest = async (request, reply) => {
-  controller.userRequires(request, "PMO-Manager-Edit-Capability", reply);
-  try {
-    const projectId = Number(request.params.id);
-    const reportDate = new Date();
-    // Get the data from the database.
-    const result = {
-      project: await projectModel.findById(projectId),
-      budget: await model.getProjectBudget(projectId),
-      status: await projectModel.findMostRecentStatusById(projectId),
-      deliverable_summaries: await model.getDeliverableSummaries(projectId),
-      change_request: await model.getChangeRequests(projectId),
-      contracts: await model.getContracts(projectId),
-      contract_summaries: await model.getContractSummary(projectId),
-      report_date: reportDate.toLocaleDateString("en-US", {
-        day: "numeric",
-        month: "numeric",
-        year: "numeric",
-      }),
-    };
-
-    const body = await getDocumentApiBody(result, "P_Budget_Report_Template.docx");
-    const pdf = await cdogs.api.post("/template/render", body, pdfConfig);
-    // Injects the pdf data into the request object.
-    request.data = pdf;
-    if (!result) {
-      reply.code(404);
-      return { message: `The ${what.single} with the specified id does not exist.` };
-    } else {
-      return result;
-    }
-  } catch (err) {
-    reply.code(500);
-    return { message: `There was a problem looking up this Project Budget Report.` };
-  }
-};
-
-/**
- * Get a specific item by ID.
- *
- * @param   {FastifyRequest} request FastifyRequest is an instance of the standard http or http2 request objects.
- * @param   {FastifyReply}   reply   FastifyReply is an instance of the standard http or http2 reply types.
- * @returns {object}
- */
 controller.getProjectQuarterlyReport = async (request, reply) => {
   controller.userRequires(request, "PMO-Manager-Edit-Capability", reply);
   try {
@@ -124,87 +79,11 @@ controller.getProjectQuarterlyReport = async (request, reply) => {
   }
 };
 
-/**
- * Get a specific item by ID.
- *
- * @param   {FastifyRequest} request FastifyRequest is an instance of the standard http or http2 request objects.
- * @param   {FastifyReply}   reply   FastifyReply is an instance of the standard http or http2 reply types.
- * @returns {object}
- */
-controller.getProjectQuarterlyReviewReportOnRequest = async (request, reply) => {
-  controller.userRequires(request, "PMO-Manager-Edit-Capability", reply);
-  try {
-    const projectId = Number(request.params.id);
-    const reportDate = new Date();
-    const fiscal_breakdown = await model.getQuarterlyFiscalSummaries(projectId);
-    // Get the data from the database.
-    const result = {
-      project: await projectModel.findById(projectId),
-      deliverables: await model.getQuarterlyDeliverables(projectId, fiscal_breakdown),
-      report_date: reportDate.toLocaleDateString("en-US", {
-        day: "numeric",
-        month: "numeric",
-        year: "numeric",
-      }),
-    };
-    const body = await getDocumentApiBody(result, "P_Quarterly_Review_Template.docx");
-    const pdf = await cdogs.api.post("/template/render", body, pdfConfig);
-    // Injects the pdf data into the request object.
-    request.data = pdf;
-    if (!result) {
-      reply.code(404);
-      return { message: `The ${what.single} with the specified id does not exist.` };
-    } else {
-      return result;
-    }
-  } catch (err) {
-    reply.code(500);
-    return { message: `There was a problem looking up this Project Status Report.` };
-  }
-};
-
 controller.getReport = async (request, reply) => {
   reply.type("application/pdf").headers({
     "Content-Disposition": 'attachment;filename="test.pdf"',
   });
   return request.data;
-};
-
-/**
- * Get a specific item by ID.
- *
- * @param   {FastifyRequest} request FastifyRequest is an instance of the standard http or http2 request objects.
- * @param   {FastifyReply}   reply   FastifyReply is an instance of the standard http or http2 reply types.
- * @returns {object}
- */
-controller.getProjectStatusReportOnRequest = async (request, reply) => {
-  controller.userRequires(request, "PMO-Manager-Edit-Capability", reply);
-  try {
-    const projectId = Number(request.params.id);
-    const reportDate = new Date();
-    // Get the data from the database.
-    const result = {
-      project: await projectModel.findById(projectId),
-      deliverables: await model.projectStatusReport(projectId),
-      milestones: await model.getMilestones(projectId),
-      alignment: await model.getStrategicAlignment(projectId),
-      status: await projectModel.findMostRecentStatusById(projectId),
-      reportDate: reportDate,
-    };
-    const body = await getDocumentApiBody(result, "P_Status_MostRecent_Template.docx");
-    const pdf = await cdogs.api.post("/template/render", body, pdfConfig);
-    // Injects the pdf data into the request object.
-    request.data = pdf;
-    if (!result) {
-      reply.code(404);
-      return { message: `The ${what.single} with the specified id does not exist.` };
-    } else {
-      return result;
-    }
-  } catch (err) {
-    reply.code(500);
-    return { message: `There was a problem looking up this Project Status Report.` };
-  }
 };
 
 /**